pyproject.toml
[project]
name = "google-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.5.0",
    "google-generativeai>=0.8.5",
    "litellm>=1.69.0",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "yfinance>=0.2.59",
]
